decoupling scrapping functions
