+++ .gitignore
@@ -1,0 +1,2 @@
+ncbi-blast-2.13.0+
+ncbi-blast-2.13.0+-x64-linux.tar.gz
